init repon

+++ src/main/java/com/example/demo/model/Echart1_dy.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class Echart1_dy {
     private String dy;
-    private Integer wantoubi;
+    private Double wantoubi;
     private Integer day;
 }
